Renamed from fft2 to fft and changed to updated helper function as adviced
