+++ ivy/functional/frontends/jax/numpy/__init__.py
@@ -1,6 +1,5 @@
 # flake8: noqa
 from . import fft
-from . import fft2
 from . import linalg
 from . import name_space_functions
 from .name_space_functions import *
