+++ ivy/functional/frontends/jax/numpy/fft.py
@@ -1,0 +1,5 @@
+import ivy
+
+
+def fft2(a, s, axes, norm):
+    return ivy.fft2(a, s, axes, norm)
