--- ivy/functional/frontends/jax/numpy/fft2.py
@@ -1,7 +1,0 @@
-import ivy
-from ivy.functional.frontends.jax.func_wrapper import to_ivy_arrays_and_back
-
-
-@to_ivy_arrays_and_back
-def fft2(a):
-    return ivy.fft2(a)
