+++ ivy_tests/test_ivy/test_frontends/test_jax/test_jax_numpy_fft.py
@@ -1,0 +1,39 @@
+from hypothesis import strategies as st
+
+import ivy
+import ivy_tests.test_ivy.helpers as helpers
+from ivy_tests.test_ivy.helpers.hypothesis_helpers.array_helpers import dtype_and_values
+from ivy_tests.test_ivy.helpers.testing_helpers import handle_frontend_test
+from ivy_tests.test_ivy.test_functional.test_core.test_linalg import (
+    _get_dtype_and_matrix,
+    _matrix_rank_helper,
+)
+
+@handle_frontend_test(
+    fn_tree="jax.numpy.fft.fft2",
+    dtype_and_x=_get_dtype_and_matrix(),
+)
+def test_jax_numpy_fft2(
+    *,
+    dtype_and_x,
+    as_variable,
+    num_positional_args,
+    native_array,
+    on_device,
+    fn_tree,
+    frontend,
+):
+    dtype, x = dtype_and_x
+    helpers.test_frontend_function(
+        input_dtypes=dtype,
+        as_variable_flags=as_variable,
+        with_out=False,
+        num_positional_args=num_positional_args,
+        native_array_flags=native_array,
+        frontend=frontend,
+        fn_tree=fn_tree,
+        on_device=on_device,
+        rtol=1e-04,
+        atol=1e-04,
+        a=x[0],
+    )
