--- ivy_tests/test_ivy/test_frontends/test_jax/test_jax_numpy_fft2.py
@@ -1,35 +1,0 @@
-from hypothesis import given
-
-import ivy_tests.test_ivy.helpers as helpers
-from ivy_tests.test_ivy.helpers import handle_cmd_line_args
-
-
-@handle_cmd_line_args
-@given(
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        large_abs_safety_factor=2,
-        small_abs_safety_factor=2,
-        safety_factor_scale="linear",
-    ),
-    num_positional_args=helpers.num_positional_args(
-        fn_name="ivy.functional.frontends.jax.numpy.fft2"
-    ),
-)
-def test_jax_numpy_fft2(
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend="jax",
-        fn_tree="numpy.fft.fft2",
-        x=x[0],
-    )
